build-obs-package.sh: Trust keys from the OBS projects
If no previous build was attempted, 'osc build' is stuck waiting for
user input to accept the OBS project keys before building the packages.
However, as stdout and stderr are being redirected to /dev/null, the
user is not aware of that rendering the whole operation suboptimal. So,
in order to facilitate the automated builds, we accept keys from the
OBS projects without asking.