common: Switch Restart reason to on-abort instead of on-failure
The mysql service could fail for numerous reason so restarting it
unconditionally does not necessarily mean that the problem will
automatically be fixed. Moreover, system configuration managers
may get confused when the ExecStartPost script is constantly running
and think that the entire service is active and fail to report the
mysql failure. As such, it's best to use a more conservative approach
when restarting the service using the 'on-abort' reason. This is also
similar to what the upstream service file is using.